requirements.txt
scipy==1.4.1
visdom==0.1.8.9
dominate==2.5.1
opencv_python==4.1.2.30
mxnet_cu102==2.0.0b20200720
termcolor==1.1.0
numpy==1.16.5
imageio==2.9.0
Pillow==7.2.0
